feat: S1 wildcard let, S4 bool text_index_contains, S5 non-mut shadowing, S6 optional const semi, S8 IO shorthand (write/write_builder/read), S10 import syntax parsing, move arg_count/arg_text to SystemEnv
- S1: let _ = expr; evaluates expr for side effects, skips binding
- S4: text_index_contains returns Bool (int64_t in C, I64 in native)
- S5: non-mut local shadowing removes old binding from scope
- S6: const semicolons optional at top level
- S8: write(text)/write_builder(b)/read() desugar at HIR level
- S10: from X import Y parses (tokens, AST, HIR, codegen no-op arms)
- arg_count/arg_text moved from SystemIO to SystemEnv effect
- codegen: SystemEnv.arg_count/arg_text dispatch to Inst::ArgCount/ArgText
- tests: 192 CLI tests pass (was 183), including new tests for all features

+++ runtime/sarif_runtime.c
@@ -1034,7 +1034,7 @@ int64_t sarif_text_index_get(void* index_ptr, uint64_t key) {
     return -1;
 }
 
-int sarif_text_index_contains(void* index_ptr, uint64_t key) {
+int64_t sarif_text_index_contains(void* index_ptr, uint64_t key) {
     SarifTextIndex* index = (SarifTextIndex*)index_ptr;
     int found = 0;
     if (index == NULL || index->entries == NULL) {
